+++ tests/test_automapper_dict_field.py
@@ -1,5 +1,5 @@
-from copy import deepcopy
 from typing import Any, Dict
+from unittest import TestCase
 
 from automapper import mapper
 
@@ -12,36 +12,68 @@ class Candy:
 
 class Shop:
     def __init__(self, products: Dict[str, Any], annual_income: int):
-        self.products: Dict[str, Any] = deepcopy(products)
+        self.products: Dict[str, Any] = products
         self.annual_income = annual_income
 
 
 class ShopPublicInfo:
     def __init__(self, products: Dict[str, Any]):
-        self.products: Dict[str, Any] = deepcopy(products)
+        self.products: Dict[str, Any] = products
 
 
-def test_map__with_dict_field():
-    products = {
-        "magazines": ["Forbes", "Time", "The New Yorker"],
-        "candies": [
-            Candy("Reese's cups", "The Hershey Company"),
-            Candy("Snickers", "Mars, Incorporated"),
-        ],
-    }
-    shop = Shop(products=products, annual_income=10000000)
+class AutomapperDictFieldTest(TestCase):
+    def setUp(self) -> None:
+        products = {
+            "magazines": ["Forbes", "Time", "The New Yorker"],
+            "candies": [
+                Candy("Reese's cups", "The Hershey Company"),
+                Candy("Snickers", "Mars, Incorporated"),
+            ],
+        }
+        self.shop = Shop(products=products, annual_income=10000000)
 
-    public_info = mapper.to(ShopPublicInfo).map(shop)
+    def test_map__with_dict_field(self):
+        public_info = mapper.to(ShopPublicInfo).map(self.shop)
 
-    assert public_info.products["magazines"] == shop.products["magazines"]
-    assert id(public_info.products["magazines"]) != id(shop.products["magazines"])
+        self.assertEqual(
+            public_info.products["magazines"], self.shop.products["magazines"]
+        )
+        self.assertNotEqual(
+            id(public_info.products["magazines"]), id(self.shop.products["magazines"])
+        )
 
-    assert public_info.products["candies"] != shop.products["candies"]
-    assert public_info.products["candies"][0] != shop.products["candies"][0]
-    assert public_info.products["candies"][1] != shop.products["candies"][1]
+        self.assertNotEqual(
+            public_info.products["candies"], self.shop.products["candies"]
+        )
+        self.assertNotEqual(
+            public_info.products["candies"][0], self.shop.products["candies"][0]
+        )
+        self.assertNotEqual(
+            public_info.products["candies"][1], self.shop.products["candies"][1]
+        )
 
-    assert public_info.products["candies"][0].name == "Reese's cups"
-    assert public_info.products["candies"][0].brand == "The Hershey Company"
+        self.assertEqual(public_info.products["candies"][0].name, "Reese's cups")
+        self.assertEqual(
+            public_info.products["candies"][0].brand, "The Hershey Company"
+        )
 
-    assert public_info.products["candies"][1].name == "Snickers"
-    assert public_info.products["candies"][1].brand == "Mars, Incorporated"
+        self.assertEqual(public_info.products["candies"][1].name, "Snickers")
+        self.assertEqual(public_info.products["candies"][1].brand, "Mars, Incorporated")
+
+    def test_map__use_deepcopy_false(self):
+        public_info_deep = mapper.to(ShopPublicInfo).map(self.shop, use_deepcopy=False)
+        public_info = mapper.to(ShopPublicInfo).map(self.shop)
+
+        self.assertIsNot(public_info.products, self.shop.products)
+        self.assertEqual(
+            public_info.products["magazines"], self.shop.products["magazines"]
+        )
+        self.assertNotEqual(
+            public_info.products["magazines"], id(self.shop.products["magazines"])
+        )
+
+        self.assertIs(public_info_deep.products, self.shop.products)
+        self.assertEqual(
+            id(public_info_deep.products["magazines"]),
+            id(self.shop.products["magazines"]),
+        )