+++ tests/test_automapper_sample.py
@@ -1,3 +1,5 @@
+from dataclasses import dataclass
+
 from automapper import mapper
 
 
@@ -18,6 +20,27 @@ class PublicUserInfoDiff:
     def __init__(self, full_name: str, profession: str):
         self.full_name = full_name
         self.profession = profession
+
+
+@dataclass
+class Address:
+    street: str
+    number: int
+    zip_code: int
+    city: str
+
+
+class PersonInfo:
+    def __init__(self, name: str, age: int, address: Address):
+        self.name = name
+        self.age = age
+        self.address = address
+
+
+class PublicPersonInfo:
+    def __init__(self, name: str, address: Address):
+        self.name = name
+        self.address = address
 
 
 def test_map__field_with_same_name():
@@ -79,3 +102,14 @@ def test_map__override_field_value_register():
         assert public_user_info.profession == "engineer"
     finally:
         mapper._mappings.clear()
+
+
+def test_map__check_deepcopy_not_applied_if_use_deepcopy_false():
+    address = Address(street="Main Street", number=1, zip_code=100001, city="Test City")
+    info = PersonInfo("John Doe", age=35, address=address)
+
+    public_info = mapper.to(PublicPersonInfo).map(info)
+    assert address is not public_info.address
+
+    public_info = mapper.to(PublicPersonInfo).map(info, use_deepcopy=False)
+    assert address is public_info.address